pyproject.toml
[project]
name = "flame-hub-client"
version = "0.2.2"
description = "HTTP client for interacting with FLAME Hub services."
authors = [
    { name = "Maximilian Jugl", email = "Maximilian.Jugl@medizin.uni-leipzig.de" }
]
readme = "README.rst"
requires-python = ">=3.10,<4"
dependencies = [
    "pydantic (>=2.10.0,<3.0.0)",
    "httpx (>=0.28.0,<0.29.0)",
    "httpcore (>=1.0.9,<2.0.0)"  # dep of httpx, contains vulnerable h11 dep so needs to be patched here
]
license = "Apache-2.0"
classifiers = [
    "Development Status :: 4 - Beta"
]

[project.urls]
homepage = "https://privateaim.de/"
repository = "https://github.com/PrivateAIM/hub-python-client"

[tool.poetry]
packages = [{ include = "flame_hub" }]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.0"
ruff = ">=0.9.0"
pre-commit = ">=4.0.0"
testcontainers = { extras = ["minio", "mysql", "rabbitmq", "redis", "vault"], version = ">=4.9.0" }
pytest-dotenv = ">=0.5.0"


[tool.poetry.group.docs.dependencies]
sphinx = "8.1.3"
sphinx-rtd-theme = "^3.0.2"
sphinx-autobuild = "^2024.10.3"
numpydoc = "^1.8.0"

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
markers = [
    "integration: mark a test as an integration test."
]
env_files = [
    ".env",
    ".env.test"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
